+++ horisapi/src/main/java/com/horis/api/model/UnidadeCurricular.java
@@ -13,23 +13,22 @@ public class UnidadeCurricular {
 
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
-	private Integer id;
+	private Long id;
 	private String nome;
 	private int cargaHoraria;
 	private String area;
-	
 	@ManyToMany
 	private List<Curso> cursos;
 	/**
 	 * @return the id
 	 */
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 	/**
 	 * @param id the id to set
 	 */
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	/**
