+++ horisapi/src/main/java/com/horis/api/repository/ProfessorRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.horis.api.model.Professor;
 
-public interface ProfessorRepository extends JpaRepository<Professor, Integer>{
+public interface ProfessorRepository extends JpaRepository<Professor, Long>{
 }
