+++ horisapi/src/main/java/com/horis/api/repository/UnidadeCurricularRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.horis.api.model.UnidadeCurricular;
 
-public interface UnidadeCurricularRepository extends JpaRepository<UnidadeCurricular, Integer> {
+public interface UnidadeCurricularRepository extends JpaRepository<UnidadeCurricular, Long> {
 
 }
